modify some names

+++ core/src/texture.hpp
@@ -6,7 +6,7 @@ namespace core {
 
 class Texture {
 public:
-  Texture(const std::string &path, size_t frame_width, size_t frame_heigth);
+  Texture(const std::string &path, float frame_width, float frame_heigth);
 
 private:
   std::string _path;
